Extended chandelay
Extends CDELAY so that modes are retained on netsplit and users are able
to join locked channels (though they will not receive chanops and +bIeik applies on join).

Race conditions are avoided by forced channel mode resynch whenever it is "created". This
is a bit noisy on the wire, and might have been problem in the 90s, but hey, it's 2013.

This permits to use much higher chandelay timeouts, in range of several days to weeks (separate
patch for increasing CDELAY).

Rationale: Solves ban evasion via split riding, solves hacking ops in places where SS/SU
is not enough (small cctld channels).

1. keep modes of empty channels, and when local user joins CDed channel, propagate those to neighbors
2. if remote user joins such empty channell, drop our chanmodes (presumably netjoin, other extcd server or expired CD)
3. there are no chanops on joining locked channel, as the channel is not in fact newly created
4. this is compatible with normal chandelay:
    * servers A and B split from the main net, let there be empty channel now where chanops left, thus is marked locked
    * user 1 joins on server A which supports extcd
    * server A sends NJOIN and propagates chanmodes to server B which does vanilla cd (thus resets chanmodes on join)
    * user 2 joins on server B (because channel is non-empty now, though with ->history)
    * user 1 quits on server A
    * user 2 quits on server B
    * server B is in its original state (modes reset and unacessible - ERR_UNAVAILRESOURCE)
    * server A is in its original state (modes kept and joinable)

Signed-off-by: sd <sd@hysteria.cz>